requirements.txt
openai
langchain
faiss-cpu
PyPDF2
streamlit
tiktoken
python-dotenv
langchain-community
pypdf